+++ api/src/main/resources/marquez/db/migration/V55.4__add_migration_lock.sql
@@ -1,0 +1,4 @@
+CREATE TABLE v55_facet_migration_lock (
+  created_at  TIMESTAMPTZ,
+  run_uuid    UUID
+);